refactor: 💡 interface move from renderer to top
