+++ src/renderer/index.ts
@@ -7,7 +7,8 @@ import { RenderNodeManager } from './elements/node/render-node'
 import Node from '../node'
 import Link from 'src/link'
 import { RenderLinkManager } from './elements/link/render-link'
-import { Transform, RendererConfigs } from './interfaces'
+import { Transform } from '../interfaces'
+import { RendererConfigs } from './interfaces'
 import { Color } from 'src/interfaces'
 import { decodeRenderId } from './utils'
