+++ src/renderer/interfaces.ts
@@ -5,12 +5,6 @@
 
 import { Color } from 'src/interfaces'
 
-export interface Transform {
-    x: number
-    y: number
-    k: number
-}
-
 export interface RendererConfigs {
     canvas: HTMLCanvasElement
     width: number
